Fix. 확장성을 위해 Carousel 로직 변경
